Fix configuration mapping when there is no prefix Argument validation in DefaultAppConfig Move annotations to a better package
