+++ archaius-core/src/main/java/com/netflix/archaius/annotations/ConfigurationSource.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.archaius.mapper.annotations;
+package com.netflix.archaius.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
